Initial commit - added src, resources, and lib directories.
The "lib" and "resources" directories are where all required libraries and resources are stored.
The "src" directory contains the engine and demo/test source code.  This engine is divided into
core, graphics, gui, and inputs.  Core/framework contains the entity component framework system
used in core/implementation.  The "inputs" directory is for reading user input to the computer,
while gui and graphics are pretty self explanitory.  The "utils" directory contains a set of
utilities used throughout the entire project.