Improved PhageSplash, fixed ParallaxRenderingActivity, and added new game.
Made PhageSplash close after loading.  Renamed DepthRenderingActivity to ParallaxRenderingActivity
and fixed bug where it was moving the objects the wrong way.  Added removeAllEntities() to
EntitySystem, added ListenerCollisionHandlerLogic, and made a function to change the gravity in
PhysicsActivity.  Also started work on a new puzzle game and commented out some shader code which
threw an error.