Fixed bug with spastic translations and rotations - however, the 360 degree rotation bug is still not fixed
The fixed bug occured because every time a transmission was received, all the deltas for all the entities
were computed.  This resulted in the interpolation transformations to be perceived as a delta from the server
when only other Entities were being updated.  These would compound rapidly, causing objects to fly off the
screen.